WWW-Authentication directories are now specified by server path and not filesystem path.  Code cleanup for www-authentication.  Major authentication enhancements to come.
Numerous other code cleanups, focusing on reducing the number of arguments passed to various functions.  The biggest change is that the query part of a URL isn't passed through all the URL-parsing functions.   #urltype.q is now unused, should remove it at some point in the future.

We update the ARG record a lot less now.


git-svn-id: https://erlyaws.svn.sourceforge.net/svnroot/erlyaws/trunk/yaws@504 9fbdc01b-0d2c-0410-bfb7-fb27d70d8b52