do not pass the --id option if the default id is used
The default id in the code is represented as the string "default". If
the --id option passes default to set the id, it becomes the atom
default which of course differs from the string "default". By not
using --id to specify the default id, we get the value "default" as
the default.

Make sure ids passed in via the --id option that yaws_sup.erl sees as
atoms are converted to lists. This matches what's done for other
options involving ids such as --check and --stop.

Other changes in this commit make sure that the $HOME/.yaws/yaws/ID
path is used consistently across the code and documentation, and also
a couple minor whitespace/indentation cleanups.