set nodelay on FCGI TCP connection, and avoid 0-length FCGI gen_tcp:recv
Set {nodelay,true} on the TCP connection to the FCGI server to improve
performance of small requests.

The code receives packets from the FCGI server and for some cases was
extracting length fields from some packets and then making further
gen_tcp:recv calls based on those length values. The code was not
checking for length values of 0 before calling recv, and passing 0 to
recv means to return all available bytes. For both correctness and
performance, the code must avoid the recv calls altogether when the
length is 0.