work around an erlang:open_port bug for CGI QUERY_STRING env var
Some versions of erlang:open_port can't handle query strings that end
with an equal sign. This is because the broken versions treat
environment variable strings ending with '=' as environment variable
names intended to be deleted from the environment, i.e. as if they
have no value. The result is that no QUERY_STRING environment variable
gets set for these cases. We work around this by appending a &
character to any query string that ends in =.

Also, clean up some comment lines that extended past 80 characters.