Use gconf keepalive_timeout for connection lifetime (Brady McCary)
Previously persistent connections were held open for a hard-coded 30
seconds, and if no request was received in that time, Yaws closed the
connection. Change this behavior to use the gconf keepalive_timeout
field to determine the amount of time the connection is held open
rather than using a hard-coded timeout. The default is still 30
seconds, but users can change the default by specifying a different
value for keepalive_timeout in their Yaws configuration.

Also, fix yaws.tex to document the keepalive_timeout setting as well
as other recent changes such as max_connections and keepalive_maxuses.