Augment yaws man page with --erlarg argument quoting info
If a multi-word argument is passed to --erlarg, it must be quoted in
the calling shell so that yaws sees it as a single argument. Add
details to the man page about this, and also show an example of how
single quotes within such a multi-word argument must receive extra
quoting.