Allow "stream processes" to close the client socket
Yaws allows application processes to take control of the client socket
via the streamcontent_from_pid directive. Sometimes during the course
of its data delivery the application might find that it needs to close
the socket, but the original design did not provide a way for the
application to tell Yaws the socket was closed. The result was that a
message like the one below would show up in the Yaws log whenever a
closed socket was returned to Yaws, due to Yaws trying to continue to
use the socket as if it were still open:

     =ERROR REPORT==== 12-May-2010::00:27:05 ===
     Yaws process died: {{badmatch,{error,einval}},
                         [{yaws,setopts,3},
                          {yaws,http_recv_request,2},
                          {yaws,do_http_get_headers,2},
                          {yaws,http_get_headers,2},
                          {yaws_server,aloop,3},
                          {yaws_server,acceptor0,2},
                          {proc_lib,init_p_do_apply,3}]}

Change yaws_api:stream_process_end/2 to allow the application to pass
the atom 'closed' as the first argument, in place of the socket
argument. This informs Yaws that the socket is closed and prevents it
from trying to keep using it, thus avoiding errors like the one above.