performance enhancements
Based on a yaws configuration consisting of an appmod registered on
"/" and simply returning {status, 204} with no response body, these
changes resulted in a 6.8% performance improvement on a modest older
Ubuntu 8.10 2.40GHz Core 2 Duo box with 2GB of RAM. The changes were
identified by analyzing fprof output, and performance was measured
using ab (ApacheBench) running on the same host as shown below:

ab -k -n 1000000 -c 1000 http://localhost:8200/

The access log was disabled and the listen backlog was set to 512, but
other than that the default configuration was used. Requests per
second went from ~14750 to ~15750. YMMV.