add configure detection of erts bits support
The websockets support relies on bits support in Erlang. The
src/Makefile used to contain a gmake shell command for checking the
erts version to see if bits support was present. This change moves
that test into the configure scripts and adds a new make variable to
include.mk that src/Makefile can use instead. This keeps configuration
checks out of the makefiles and in the configure script where they
belong.

Also fixed the test for bits support, which would have been incorrect
once the erts major version number increases above 5 because it also
unconditionally wanted the erts minor number to be greater than 6, so
a major/minor version such as 6.0 would have failed.