filter erl output for setting ERLDIR
A user reported a problem with yaws-1.89 where ERLDIR was getting set
incorrectly because his R14B01 Erlang installation was emitting the
string "Running Erlang" as part of the output used to set the ERLDIR
variable. (I don't know why, since that string doesn't appear anywhere
in the sources as far as I can see.) Change the command used to set
ERLDIR so that only the last line of output is kept.