add test target to top-level Makefile
Add test target to top-level Makefile to allow tests to be built and
run from that level.

Also fix the start target in test/support/include.mk.in to only try to
start a yaws instance if the current working directory contains a
yaws.conf file. The lack of this check was preventing half the tests
from running under "make test".