do not use ERL_DRV_USE when de-selecting sendfile socket
The sendfile driver does not own the HTTP sockets to which it sends
data, and so it never closes them. The previous commit erroneously set
ERL_DRV_USE on the driver_select(..., 0) call, which causes erts to
justifiably complain about the driver not implementing the stop_select
call. Remove ERL_DRV_USE for this case. Thanks to Tuncer Ayaz.