Figuring out proper yaws_session_server backend.
This fixes the problem with getting hold of the yaws_session_server
backend when running in embedded mode. In this mode, no #gconf{}
record may be available when the yaws_session_server want to figure
out what backend to use. The fix stores the 'pending' configuration
as a yaws appplication environment variable 'embedded_conf', which
makes it possible for the yaws_session_server to get hold of the
backend when starting up. If still no #gconf{} can be obtained,
we will fallback to the default yaws_session_server backend.