fix yaws_server dialyzer issues
In yaws_server the old function ut_open used to return either {bin,
Bin} or an open file handle depending on how it was called. A number
of other areas of the yaws_server code also handled either a file
handle or {bin, Bin}. The ut_open function was changed in June 2002 to
deal only with binaries, and dialyzer correctly pointed out that a
number of areas of the code were obsolete as a result. Change
yaws_server to get rid of ut_read and ut_close, then rename ut_open to
ut_read and have it just return a plain binary. Remove all vestiges of
code relying on the old behavior related to open file handles.

This clears up all known dialyzer issues in yaws_server.