add status option to page options (capflam)
Scripts can return {page, {Options, Page}} to make Yaws return a
different page than the one being requested, where Options is a deep
list of options. Previously the only type of option was {header, H}.
This patch adds the type {status, Code} to set the HTTP status code of
the response.

It also delays the accumulation of these headers to preserve it during
the handling of the request on Page. Both 'delayed_regular' and
'delayed_directory' need fewer checks than other requests because
they're used on the second pass and many tests have already been done
during the first pass.