allow hard reload of conf without stopping (capflam)
Allow a hard reload of the configuration when the server is running,
without stopping the application. This happens when the configuration
has changed too much (e.g. the IP address of virtual server). This is
done by stopping all gserv processes and restarting it with the new
configuration, so it is pretty brutal.