add control for "Expires" and "Cache-Control" response headers (capflam)
Allow an expires list in the server configuration and add "Expires"
and "Cache-Control" headers in responses if the mime type of the
document matches an element of this list. Like the Apache mod_expires
module, the "expires" configuration directive controls the setting of
the Expires HTTP header and the max-age directive of the Cache-Control
HTTP header in server responses. The expiration date can be relative
to either the time the source file was last modified or to the time of
the client access. It is possible to have multiple expires directives
for a virtual server.

Here is an example:

  <server www.yakaz.com >
    ...
    expires = <image/gif, access+2592000> <image/png, access+2592000>
    expires = <image/jpeg, access+2592000> <text/css, access+2592000>
    expires = <application/javascript, modify+2592000>
    ...
  </server>