support external handler to interpret php scripts (capflam)
phpfcgi is now deprecated. Instead, you must use php_handler. This
directive is used to customize the handler to interpret .php files. It
can be one of the following definitions:

  - php_handler = <cgi, Filename>
  - php_handler = <fcgi, Host:Port>
  - php_handler = <extern, Module:Function | Node:Module:Function>

If there is no php_handler or deprecated phpfcgi directive, we use the
value of php_exe_path. Its default value is "/usr/bin/php-cgi" but it
can be customized.