server-specific logger_mod and auth_mod (capflam)
Make logger_mod and auth_mod configuration variables local to each
virtual server. The logger_mod variable can be used to customize
access and auth messages. The auth_log variable previously was global,
but it now lives in the server configuration and the global variable
is deprecated.

The new behaviour yaws_logger must be used to define external modules
to log messages.