always store a tuple in #headers.authorization (capflam)
The function yaws:parse_auth/1 returned "undefined" if it couldn't
parse the value of the "Authorization" header. In this case, we lost
the original value of this header.

Always store a tuple of the form "{User, Pass, Orig}" even if the
value can't be parsed. In this case, "User" and "Pass" are set to
"undefined".

Also, export yaws:parse_auth/1 so that 3rd-party modules can set