increase default maximum wait-started time
Increase the default maximum time for the yaws --wait-started option
from 6 seconds to 30 seconds. The test suite was seeing random
failures due to 6 second wait not being enough. (The option doesn't
mean it waits for 30 seconds every time, just that it will wait for a
maximum of 30 seconds for yaws with the specified id to start.)