use {wordsize, external} in configure script
To properly check to see whether the Erlang emulator is 32-bit or
64-bit in the configure script, call erlang:system_info with the
argument {wordsize, external} instead of just wordsize. The latter
reveals only the internal word size, which happened to work previously
(and is the only option for older versions of Erlang) but doesn't work
for halfword emulators. Fallback to an argument of just wordsize for
older versions of Erlang.