let configure handle erlang built from git clone
A normal Erlang installation has versioned directory names for
applications such as sasl, kernel, and ssl. The yaws configure script
checks for these. But an Erlang built from a git clone of the OTP
repository, the versions for these applications do not exist. Trying
to build yaws using erl built in such a git clone would fail due to
this lack of versioned applications.

Change configure.in to look for normal application versions first and
if not found, look for application dirs containing vsn.mk files. If
found, use those versions instead. Also change configure.in to get the
erts version from running the emulator instead of from the erts
versioned directory name.