fix reverse proxy problem (issue #60)
Change yaws_server.erl to not exit on {'EXIT', From, Reason} messages
found when erasing transient messages, a change originally introduced
in commit c0755739. Doing so was causing reverse proxy to no longer
work, as reported in issue #60 on github and as experienced by several
users. Note however that this might be a temporary fix, since it might
be better to figure out the source of the EXIT message, but I'm
committing and pushing this fix now for the benefit of those waiting
to use the reverse proxy feature.

Thanks to Manuel Durán Aguete for providing a test case making it easy
to reproduce the problem and test the fix.