Use filter to reduce the amount of trace messages logged by Yaws
When traffic or http tracing is enabled, a filter can be defined to
disable trace logging for specific requests. When Yaws receive a request,
if tracing is enabled and the filter is evaluated to true, trace messages
are logged. Else tracing is disable for this request.

Filter is defined using a syntax similar to the condition part of match
specifications. We can match on the client's ip address or request headers.

For example, to trace requests coming from 127.0.0.1 on all gif images,
we can do:

  1> yaws_trace:set_filter({'and',
                            {equal, ip, "127.0.0.1"},
                            {match, {request, path}, "^.*\.gif$"}}).
  ok