allow functions to be specified as values in ehtml
Ehtml now allows functions to be used as values. You can specify a
function value as an anonymous arity 0 function, or as a tuple
consisting of a module, function name, and argument list (i.e.,
{M,F,[Args]}). Functions must return valid ehtml values, and they may
return other functions as long as those functions eventually return a
valid non-function ehtml value.

Augment yaws.tex to document this new feature. Add a new eunit test
for the feature as well.