add config setting for acceptor pool size
Add new config setting to allow the size of the acceptor process pool
to be set to something other than the default. The default size is the
same as what it was prior to this change. Also add documentation for
the new setting, and augment the yaws.conf.template with information
about it.