Finalise support for multiple and partial frames in a packet.
With this commit, Autobahn cases 1.* and 2.* pass.

This means when multiple websocket frames are in a single packet,
all of them will be unframed and returned to the endpoint in a list.
If the last frame in a packet is incomplete, exactly as many bytes as
are needed to complete the frame will be read from the socket.

To achieve this in active mode, the socket is actually in {active, once}
mode. More data is read using gen_tcp:recv, and the socket is returned to
{active, once} mode before returning. Any data left on the socket
when changing back to active mode is sent to the owner process
in an erlang message.

Frequently (proportonal to the number of frames) changing the active
property of sockets might be slow, I haven't benchmarked this. For now
this is tested only using {active, once}.