Add reltool based release support
Running "rebar generate" now creates a self-contained yaws system
under the build directory's "rel" subdirectory. The yaws script it
provides at ./rel/yaws/bin/yaws isn't the same as the normal yaws
script (the one normally found at ./bin/yaws); rather, it's a special
script that starts yaws and the Erlang applications on which it
depends as a local self-contained Erlang node. You can run

./rel/yaws/bin/yaws console

to start an interactive yaws node, or

./rel/yaws/bin/yaws start

to run it as a daemon, which you can later stop with

./rel/yaws/bin/yaws stop

This script does not accept the command-line options that ./bin/yaws
does, at least for now. If this is a problem, please raise an issue at
the Yaws github repo or on the Yaws mailing list.