Add options to configure deflate compression behaviour
By adding "<deflate> ... </deflate>" structure in vhosts configuration,
it is possible to configure how deflate compression will be applied
and when it will come into effect. Now we can:

 * define the smallest response size that will be compressed
 * define the compression level to be used
 * specify the zlib compression window size
 * specify how much memory should be allocated for the internal
   compression state
 * choose the strategy used to tune the compression algorithm

All these parameters are used when a zlib stream is initialized for
compression.

It is also possible to define all compressible mime types.
Here is an example:

<server localhost>
  ...
  deflate = true
  <deflate>
    min_compress_size = 4096
    compression_level = best_compression
    mime_types        = defaults, image/*
    mime_types        = application/xml, application/xhtml+xml, application/rss+xml
    mem_level         = 9
    strategy          = default
    window_size       = 15
  </deflate>
  ...
</server>