report uncaught exception as server error 500
In yaws_server we used to catch exceptions thrown by application code and
ignore them. Now, catch them and report a 500 error back to the client, and
include an error message similar to what occurs for unexpected exits in
application code. This makes debugging applications that unexpectedly throw
exceptions much easier. Thanks to Richard Carlsson and Samuel Rivas for
reporting this.