Add support for precompressed static files
By setting use_gzip_static to true in deflate options, in a vhost
configuration, It is possible to serve precompressed versions of
static files. Yaws will look for precompressed files in the same
location as original files that end in ".gz".

Only files that do not fit in the cache are concerned and the mtime
of a precompressed file must be higher than the one of original file.