add yaws_api:get_listen_port/1
For testing scenarios it can be useful to configure yaws to run on an
ephemeral port by setting the server port number to 0. Add the
yaws_api:get_listen_port/1 function to enable callers to retrieve the
actual port number of a server from its listen socket. The single argument
to the function is the #sconf{} record for the target server.