Keep the original request in #arg{} and check it against auth rules
Because the request can be rewritten, it could be useful to backup the
original one. Appmods or Yaws scripts could access it by reading the field
orig_reg in #arg{}. For CGI scripts, the variable REQUEST_URI is now built
using the original request (SCRIPT_NAME, PATH_INFO and QUERY_STRING could
be used to know the info about the rewritten request, if needed).

The original request is also checked against the authorization rules. So, if
the request is rewritten, Yaws will check the original request and the
rewritten one. Access must be granted for both urls to accept the request.