changes for R16B01
In R16B01, crypto:sha/1 is deprecated in favor of crypto:hash/2, but the
latter is not available prior to R15B02. Change the configure script and
rebar support to look for crypto:hash/2 and use it if found, otherwise fall
back to crypto:sha/1.

Also in R16B01 the public_key application, used by some of the tests, now
requires the asn1 application to be started first. Change the tests to
accommodate that.