Use Yaws linked-in driver by default to send large file
In a previous commit (c887a374), the use of sendfile (OTP and Yaws versions)
was disabled. The OTP version is buggy, so we recommand to not use
it. But the Yaws linked-in driver can safely be used. So, there is no reason
to not use it. Now, it is the default method.