combine multiple Accept-Encoding headers
According to HTTP 1.1 section 4.2, a receiver may combine multiple
instances of any header whose value can be a comma-separated list, which
applies to the case of receiving multiple Accept-Encoding headers. HTTPbis
in section 3.2.2 keeps the same rule. Yaws was using only the first such
header and ignoring any others, but should instead be combining the
multiple Accept-Encoding headers into one before processing the combined
value. This commit fixes that and adds a new test case for it.