Relax rules that allow the parsing of a query string or an urlencoded body
The parsing of a query string or an urlencoded body was allowed, respectively,
for 'GET' or 'POST' requests. Now, this can be done for any kind of requests. Of
course, a body is parsed only if its Content-Type is set to
'x-www-form-urlencoded'.

In the same way, a body can now be parsed as a multipart content if its
Content-Type is set to 'multipart/form-data', regardless of the request type.

As a consequence of these changes, yaws_api:getvar/2 now returns the merged
result of values found in the query string and in the urlencoded body (if any).