Add support for binary and unicode FCGI params
Allow names and values in FCGI extra_env parameter lists to be either
lists or binaries, and handle them properly as unicode as well. Fix
the documentation for extra_env in yaws_api.5. Change the comment
describing extra_env in yaws_api.erl to note that binaries are allowed
too.