Fix ybed embedded supervision example
Modify the code for the ybed supervision example to link the ybed
child to its supervisor, specify the temporary restart strategy for
that child, and have the child exit normally after it starts the
necessary Yaws children in its supervisor. We use temporary rather
than permanent because the child needs to run only once, whenever the
supervisor starts.

Thanks to Lyn Headley for tracking down these issues and suggesting
the fixes described here.