Remove very old use_old_ssl directive
This directive was added to use old SSL implementation. It was removed in
Erlang/OTP R15. So, because Yaws works with releases upper to R14B02, this
option is only supported with Erlang/OTP R14B02, R14B03 and R14B04. For all
other supported releases, Yaws will fails is use_old_ssl is set to true.

To be honest, if you use SSL in Yaws, you'd better upgrade your Erlang/OTP
release. The new SSL implementation is far better than the old one and many bugs
are fixed in recent releases.