Move username/password tuples from server state to an ETS table
Previously, basic authentication in YAWS stored {Username, Password}
tuples as a list in the server state. By default, a crash in YAWS
generates a crash dump which is served over HTTP, thus remotely
exposing all passwords and usernames in plaintext.

This patch moves the {Username, Password} to an ETS table so that
sensitive information is kept on the server rather than leaked.