Fix a infinite loop when a client sends a request on "." or "./some-path"
When a client sends a request on "." or "./some-path", Yaws will loop
infinitly. This is reproducable by using a catch-all appmod (on "/")

The loop can occur in "yaws_server:is_revproxy/3",
"yaws_server:is_redirect_map/2" or "yaws_server:filter_auths/2", depending on
the configuration. In these functions there is a check on the path to return
when it is equal to "/". We must do the same when the path is ".".

This patch fixes #271.