Refactor the testsuite to use common_test instead of a in-house framework
The testsuite has been rewritten to use Erlang common_test framework.
Internally, eunit is used to check results. The TS can be launched with the
'make check' command. It is only available with the autotools.

testsuite/README.md describes how the testsuite works, how to launch specific
suites or testcases and how to write new tests.

ibrowse dependency has been removed. So it is possible to run the TS without
connectivity. We use a custom HTTP client to do checks. So it is easy to adapt
it when needed. The API to write new tests is in the module
testsuite/testsuite.erl.