Check "Transfer-Encoding" header first, before "Content-Length" header
Usually, when "Transfer-Encoding" header is set to "chunk", "Content-Length"
header is not set. So it is safe to check these headers in any order. In Yaws,
"Content-Length" header was checked first.
But, some clients (e.g. inets) set the "Content-Length" header to 0 when the
body is chunked. So, in this situation, the body was skipped by Yaws. And with
persistant connections, subsequent requests was rejected.

So, to be compliant with these buggy clients, we check "Transfer-Encoding"
header first to know if the body is chunked. Then, if not, we check
"Content-Length" header to know the body size.