Rework how random ports are handled
It is possible to let Yaws choose the listening port of a server by leaving it
to 0. Now, this port is chosen during the configuration parsing. There are 2
reasons to do it this way:

  * First, the #sconf record has the real listening port instead of 0. This
    avoids the checks to get it. This is also convenient to have it when the
    configuration is retrieved or dumped.

  * Then, if there are 2 or more servers configured to use a random port, each
    one will have its own port. Each server will be isolated in its own group.
    So it is possible to have in same time a HTTP and a HTTPS vhost using a
    random port.

There is a drawback when the configuration is reloaded. All servers will be
restarted and will use a new listening port. But random ports are used for test
purposes so this is not really an issue.

We use IANA range for dynamic or private ports (49152 -> 65535) to choose a
port. We try to find the first free port in this range. An environment variable
is used to know the last assigned random port (attached to the kernel
application, so it is possible to reserve ports manually before Yaws is
started).

This new feature is mainly used by the new testsuite.