Fix #283: FastCGI protocol error for empty HTTP body requests
In my setup I receive the following 500 error when POSTing an empty
request to any PHP script.

  CGI failure: {"recv from application server failed",closed}

My investigations showed that the FCGI_STDIN <<>> frame was sent twice
and the FastCGI server closes the connection right away, (most of the
time) before having time to anwser.

The FastCGI protocol describes the end of a stream as an empty record
for that stream. What Yaws does is it basically closes the stream
twice. Moreover, a stream is described as zero or more non-empty
records, followed by an empty record.

http://www.mit.edu/~yandros/doc/specs/fcgi-spec.html#S3.3

  "A stream record is part of a stream, i.e. a series of zero or more
  non-empty records (length != 0) of the stream type, followed by an
  empty record (length == 0) of the stream type."

http://www.nongnu.org/fastcgi/#id2854157

  "Furthermore, it will forward any POST data it receives from the
  client to the application via the STDIN stream. Once all data have
  been forwarded, an empty STDIN packet is sent to close the stream."

Though the specification is not as precise as an RFC, sending two
empty records is not in the spirit of the spec.

Add special handling for empty body requests to
yaws_cgi:fcgi_pass_through_client_data/1.