Upload build artifacts producted by Travis on external repo
Because Travis logs are not suffisant to find bugs when the testsuites execution
fail, we upload build artifcats on an extern repository.

For now, this is done on github, in the repository
capflam/yaws-travis-build-artifacts. There is no versionning, we just create new
release for each travis build and add assets in its release.