Fix a leak when a server conf is updated because of a forgotten ets table ref
Each server owns an ets table to store some information about files and scripts
and to cache small static files. When the server configuration is updated, the
old #sconf record is replaced by a new one and a new ets table is created. But
the old one was never deleted. So we could have a proliferation of forgotten ets
tables leading to a memoring leak after successive updates.

So now, instead of creating a new ets table when a server configuration is
updated, we keep the old one. This avoids useless creation of ets tables and
fixes the memory leak.