Add support of hashed passwords in yaws configuration and in .yaws_auth files
Now, it is possible to used hashed password to define user's credential. In the
yaws configuration, you can add following lines is <auth> sections:

  user = "User:{Algo}Base64Hash"

and in .yaws_auth files:

  {"User", "Algo", "Base64Hash"}.

Algo is one of: md5 | ripemd160 | sha | sha224 | sha256 | sha384 | sha512

and Base64Hash is the result of the hash functions, encoded in base64. For
example:

  "/N4rLtula/QIYB+3If6bXDONEO5CnqBPrlURto+/j7k=" for "bar" password.

yaws startup script has been updated to add a way to generate user's credential:

  $> yaws --auth --algo sha256 USER
  baz

  User's credential successfully generated:
      Put this line in your Yaws config (in <auth> section): user = "USER:{sha256}uqWglk0zIPvAxqkiFARTyFE+okq4/QV3A0gEqWckgJY="

      Or in a .yaws_auth file: {"USER", "sha256", "uqWglk0zIPvAxqkiFARTyFE+okq4/QV3A0gEqWckgJY="}.