Improve finding of private ports
Since df1d7bf Yaws does not really open port 0 anymore, but tries to
open one from 49152 onwards. This leads to "badbind" errors when e.g.
executing multiple separate common test runs which use Yaws repeatedly.

With this patch port 0 is opened instead, the port number retrieved, and
closed again. This still is not completely safe, as the port could be
taken by another application in the short time Yaws tries to listen on
it again, but in practice I could not reproduce badbind errors.

This approach leaves all advantages of df1d7bf, but also respects the
local range for dynamic ports if specified, see
http://stackoverflow.com/a/924337.

Here is a test program which illustrates the problem. With this patch
it runs continuously, without it stops with "badbind":

    #!/bin/bash

    NUM_CT_CONCURRENT=4

    trap "exit" INT TERM
    trap "kill 0" EXIT

    sim_testcase () {
        echo "ct run $1, testcase $2"
        output=`erl -pa ebin -noshell \
            -eval 'yaws:start_embedded("www/testdir", [{port, 0}], [], "'$1'")' \
            -eval 'timer:sleep(1000)' \
            -eval 'yaws:stop()' \
            -eval 'init:stop()' \
            2>&1`
        retval=$?
        if [[ $retval -ne 0 ]] ; then
            case $output in
                *badbind*) echo "---> BADBIND <---" ;;
                *) echo "$output" ;;
            esac
        fi
        return $retval
    }

    sim_ct_run () {
        n=$(($2+1))
        if sim_testcase $1 $n ; then
            sim_ct_run $1 $n
        else
            kill 0
        fi
    }

    set -m
    (
        for tag in `seq 1 $NUM_CT_CONCURRENT` ; do
            sim_ct_run $tag 0 &
        done
        wait
    )