Rewrite Yaws website
This is a huge rewritting to have a better look and feel. In addition, the
release process has been slightly changed to automatically generate and commit
the news and the changelog corresponding to the new release. The list of all
github contributors is also automatically updated.

Note: README.release has been added to explain how to create a new release.