Get the port number from the Host header in yaws_api:request_url/1
yaws_api:request_url/1 returns the parsed url (a #url{} record) that the
client requested. #url.host field is get from the Host header, with a fallback
on #sconf.servername. But #url.port was always retrived from the #sconf record
without any checks on the Host header.

With this patch, we try to extract the port number from the Host header and we
fallback on the #sonf record only if there is no Host header. This way,
yaws_api:request_url/1 is now returning the expected result.

Fix #297. Thanks to Thomas Rackwitz for reporting the issue and working on it.