Changes for Erlang/OTP 20.0
In testsuite.erl, explicitly disable SNI if not set because in
Erlang/OTP 20.0, it is set by default to the Host value used in
ssl:connect/4 (Christopher).

Fix old-style catch in yaws_server.erl that caused a compilation
warning under Erlang/OTP 20.0 (Steve).

Remove an old zlib workaround in yaws_zlib.erl, as it's no longer
needed (Christopher).

Add versions 19.3 and 20.0 to the .travis file.