Fix #322: Return multipart upload file path for errors
Add the return_error_file_path option for the function
yaws_multipart:read_multipart_form/2. When this option is supplied,
Yaws returns the path of the temporary upload file as part of an error
tuple if errors occur when opening or writing the file, or if the
upload size exceeds the maximum upload file size specified with the
max_file_size option. This allows the caller to delete the temporary
file or perform other custom cleanup.

Always return a tuple as the second element of the error tuple
returned from yaws_multipart:read_multipart_form/2 to allow the temp
upload file path to be returned when requested.

Add a description for return_error_file_path to the Yaws documentation.

Add a new unit test for return_error_file_path.