Add the ysession_cookiegen configuration variable to allow users to specify a module that can generate session cookies for yaws_session_server. Add handling of this variable to yaws_config. Document the ysession_cookiegen configuration variable in yaws.tex, yaws.conf.5, and yaws_api.5.
Add the ysession_cookiegen field to the gconf record. Use it in
yaws_session_server when it needs to generate a session cookie. If the
field specifies a module, call the module's new_cookie/0 function to
get a new session cookie. Add setter and getter functions for this
field to yaws.erl, and also modify setup_gconf to setup this field and
other missing ysession fields.

Add a new test to yaws_sessions_server_SUITE to verify that
yaws_session_server properly calls the custom session cookie generator
module when configured.