Fix #332: recognize '?' in conf strings
The documentation for serveralias states that it allows wildcards,
specifically that "'?' matches one character unless that character is
a period ('.')", but issue #332 shows that using a '?' as specified
results in a bad conf error.

Modify yaws_config to allow '?' characters in configuration
strings. Modify the yaws_sni_strict test configuration to use '?' in a
serveralias configuration setting as a regression test.