Allow config-specified extra response headers
Enhance server configuration to allow an optional block of extra
response headers to be specified. Extra headers to be added to
responses can be specified explicitly, headers can be specified to be
erased from responses, and modules to handle extra response headers
can be specified as well.

Add support to yaws:setup_sconf/2 for programmatically configuring
extra_response_headers.

Add documentation for extra response headers to the TeX document and
the yaws.conf.5 man page.

Add a new unit test for extra response headers.

Refactor yaws:erase_header/1 to work not just on a certain set of atom
headers, but on all headers whether they're atoms, binaries, or lists,
and regardless of whether they're fixed fields of the #outh{} record
or they're stored in the #outh.other field.