Fix #376: teach bindings to remove undefined variables
When {bindings, [{Key,Value},...]} is used, anything that looks like a
variable in returned out/1 content that isn't specified as a key is
left as is. This can result in pages containing content such as the
following, which can be surprising and incorrect:

    <p> %%KEY%% </p>

Add strip_undefined as a special value that developers can add to the
bindings list to cause the bindings mechanism to delete the text of
anything in returned out/1 content that looks like a variable but is
not defined with a value. In the above example, if KEY is not defined
and strip_undefined is specified, the resulting page would have the
%%KEY%% text stripped out and so the result would instead be:

    <p>  </p>

Allow strip_undefined in ssi binding lists as well, for consistency,
but the default behavior there is already to strip out undefined
variables, and so for ssi specifying strip_undefined does not change
the result.

To allow all {bindings, [...]} directives returned within a given
virtual server to strip undefined bindings, add the boolean server
configuration variable strip_undefined_bindings. Setting it to true
has the same effect as including strip_undefined in all {bindings,
... } directives returned within that virtual server. It's false by
default to preserve backwards compatibility.

Add documentation for strip_undefined and strip_undefined_bindings.

Add new tests that apply strip_undefined and strip_undefined_bindings.