Add support for extra response headers module
Note that this is an experimental change -- no guarantee it will ever
make it to master in this form.

Add support for programmatically configuring extra_response_headers to
yaws:setup_sconf/2.

Add support for the <extramod Module> directive within a server
configuration <extra_response_headers> block. The specified Module is
expected to be a module exporting an extra_response_headers/1 function
taking an #outh{} record as an argument and returning a list [{Header,
Value}] specifying extra response headers, where both Header and Value
are expected to be of either string or binary type. The #outh{} record
specifies the headers that will be returned if the
extra_response_headers/1 function returns an empty list.

There are no documentation changes in this commit, which means the
docs are out of sync with the code.