Fix #377: try to preserve response header order
RFC 7230 section 3.3.2 discusses header order, and its opening
statement says that header order is not significant. But it then says
that it's good practice to order general headers and control headers
first. Yaws typically follows this good practice through its use of
its #outh record, which contains "known" common headers and an
"others" list for uncommon or unknown headers.

Change the handling of #outh.other so that the order of headers
returned from user applications is not reversed. Add a test to verify
this order and to verify that #outh.other headers do not precede
common headers known in the #outh record.