Fix some feature handling in yaws_dynopts.erl
The handling of SSL SNI, SSL handshake, and whether to start error
logging in yaws_dynopts was incorrect. The macros for these features
were always defined to either true or false for the generated
yaws_dynopts, and so it would try to use the features even for
releases that didn't support them. This would result in seemingly
random testsuite failures. Fix the handling of these features so that
their respective macros are defined only when the features are
present.