Make parse_query caching smarter
The yaws_api:parse_query/1 function caches query parsing results in
the process dictionary, but this returned incorrect results if the
same process tried to parse two or more different queries. Cache the
original query string along with the result of parsing it so that the
result is reused only if the query string matches. Add a new test to
yaws_api_SUITE for this case.