Fix #384: disallow multiple Content-Length headers
As per RFC 7230 section 3.3.3, multiple Content-Length headers with
different values are not allowed. Detect this condition in
yaws:http_collect_headers and return status code 400 if found.

Add two new unit tests: one that checks for multiple different
Content-Length headers and verifies it causes a 400 status code, and
another that verifies that multiple identical Content-Length headers
are not treated as an error but are instead replaced with a single
header.