Fix #388: add appmod_name field to #arg record
The appmod_name field makes it possible for the extra_response_header
callback to know in which appmod context it is executing and hence be
able to return the proper set of extra HTTP headers. For requests
handled by something other than an appmod, the appmod_name field has
the value undefined.

Extend a few unit tests, including the extra response headers test, to
verify the appmod_name field.

Update the #arg record shown in the Yaws API man page and in the LaTeX
documentation. Describe the appmod_name field in the documentation.