Update github repository URL
The Yaws github repository URL has changed to

https://github.com/erlyaws/yaws

Change occurrences of the old URL to the new one.

Note that the file rebar-templates/yawsnode.rebar.config still has an
occurrence of the old URL; this is because this file will be going
away soon when the rebar3-support branch merges to master.