Add yaws_api:reformat_header/3
To allow header formatting functions to receive normalized data, add
yaws_api:reformat_header/3, which is just like reformat_header/2 but
takes as a 3rd argument a list of data format options for header and
value data passed to the formatting function. The list is expected to
contain the atoms string or binary. Only the final item in the list
matters (it overrides values earlier in the list), and if it's
anything other than string or binary, it disables data normalization.

Augment the headers_suite reformat_header test to verify
reformat_header/3.

Document reformat_header/3 in the yaws_api.5 man page.