Change option handling for in reformat_header/3
Rather than making later options take precedence over earlier ones,
use the first option. This makes it easier for a caller to override a
list by prepending a new option to the front of the list.

Also add a reformat_header/3 clause taking a single atom as its 3rd
argument.

Adjust the documentation to account for both changes.