Do not try to fetch messages via process_info/1
Since October 2018, the `erlang:process_info/1` does not retrieve
messages any longer. Therefore, this code seems to be a no-op.
One may argue it could even be confusing to see in the debug dump a
number of messages > 0 but no messages. We could explicitly retrieve
the list of messages via `erlang:process_info/2` but that could
potentially be unsafe.

References:

https://github.com/erlang/otp/pull/1745