Add OTP 23.0 and drop 17.x
Drop support for 17.x and set the minimum supported OTP version to
18.0. Change references to 17.0 to 18.0 in documentation and HTML
pages.

In OTP 23 we get a warning for the deprecated ssl:cipher_suites/0
function. For now, turn off -Werror by default; it will be reenabled
once the warning is fixed. If you want to build with warnings turned
on, use "make all WARNINGS_AS_ERRORS=-Werror".

Add 23.0 to the Travis-CI YAML file, and remove 17.x from it.