Fix #413: Add WebAssembly MIME type
Add application/wasm for WebAssembly files. This is not officially
standardized yet, but at https://www.w3.org/TR/wasm-web-api-1/ the
WebAssembly API documentation specifies it, and a number of tools
already use it.

The last mime.types update was Aug 2019, so perform a new update.
These updates do not include commented-out MIME types, and as a result
in this case the update is tiny.