Make response for OPTIONS * configurable
Previously the response to OPTIONS * was hardcoded to send HTTP 200 and
the header Allow: GET, HEAD, OPTIONS, PUT, POST, DELETE.

This change adds the options_asterisk_methods setting in the sconf for a
list of custom HTTP methods allowed for OPTIONS *. By default the list
is empty, and Yaws sends an HTTP 400 Bad Request. If
options_asterisk_methods contains HTTP methods, Yaws responds to
OPTIONS * with HTTP 200 and an Allow header with the contents.

An additional knee-jerk fix was to add the Date header to
yaws_server:deliver_xxx/5. According to RFC 7321, Section 7.1.1.2. Date,
this header MAY be sent for 1xx and 5xx responses and MUST be sent
otherwise.