Add convenience function json2:obj_fetch/3
The json2:obj_fetch/2 either returns an object member's value or exits
with a struct_no_key error if the object member is not found. Add
json2:obj_fetch/3 that returns either the object member's value or a
default value if the object member is not found.

Add a new unit test for json2:obj_fetch/2,3 to the json2_SUITE.

Also in json2_SUITE, fix the test source directories for the
json2_decode_valid and json2_decode_invalid tests. Add asserts to
verify that one or more test files are processed by those tests.