Detect externally rotated logfiles
Currently, Yaws does not play well with an external log rotation
mechanism (Linux logrotate or BSD newsyslog). When such an external
program rotates a Yaws log, it will customarily move the existing log
to a new name, open a new logfile with the old name, and (optionally)
send a (configurable) signal to the program that writes to the log, to
notify it to re-open the logfile.

Yaws supports wrapping its logs at a given size (configuration
variable log_wrap_size), but it does not support wrapping it in a
time-based fashion (e.g., wrap once a day at midnight), which is why
an external log rotator is useful.

The problem is that in such a case, Yaws will continue writing to the
old logfile. There is logic to detect that the logfile has been
wrapped, but it does not work if the log rotator creates a new file
in place of the old. In such case, Yaws should detect that the logfile
it sees at the expected path is smaller than what it knows to have
written to the log, and take that as a signal that the log has been
wrapped.

This patch adds support for just this. To make use of it, one should
arrange for the log rotator to issue a 'yaws --hup' post rotating the
files. This will trigger Yaws to immediately re-evaluate the
conditions and lead to migrating to the newly opened logfile.
Otherwise, it will take up to 10 minutes for this to happen, and Yaws
will have been writing to the old file (via the file descriptor it is
holding to) prior to that.