Add case clause for '*' target in reformat_request/1
This commit adds a case clause in yaws_api:reformat_request/1 to handle
the * request target.

If e.g. an OPTIONS * request is sent when trace is enabled, the call
chain ends up in yaws_api:reformat_request/1 which crashes with case
clause. This is because '*' is the path and this case is not handled.