Use special logger handler for report.log
Use logger instead of error_logger for report.log. This makes it
possible to run yaws when no error_logger process running, since we
don't try to install custom handlers in it.

Because of logger_std_h differences in Erlang/OTP 21.0 through 21.2
that prevent these changes from working correctly for those versions,
raise the minimum Erlang/OTP version that Yaws supports from 21.0 to
21.3. Anyone needing 21.0-21.2 support can use Yaws release 2.1.0.
Change all places mentioning 21.0 as the minimum to be 21.3 instead.