Fix #440: handle TLSv1.3-only server
The Erlang/OTP ssl configuration settings "secure_renegotiate" and
"client_renegotiation" don't apply to TLS version 1.3, but Yaws sets
both to defaults that differ from Erlang/OTP defaults and so always
tries to include them when setting up SSL listening, which results in
an error for servers configured for only TLSv1.3.

If a server is configured for TLSv1.3 only, set secure_renegotiate and
client_renegotiation to undefined. Also modify supported configuration
settings for these variables to allow them to be set to true, false,
or undefined, where the latter setting restores the Erlang/OTP
default. Modify the documentation to describe the undefined setting.

Add a new test to sconf_SUITE to verify settings in an #ssl{} record
when a server is configured for only TLSv1.3. Note that this test is
skipped for OTP 21.3 since it doesn't support TLSv1.3.