Fix handling of SOURCE_DATE_EPOCH
If no SOURCE_DATE_EPOCH is specified via configure, then it can't be
set as an environment variable in doc/Makefile.am build rules. Change
the build rules to check if SOURCE_DATE_EPOCH is non-empty and use it
if so, otherwise leave it out of the commands.

Also fix verbosity for make in doc/Makefile.am. If V=1 is passed on
the make command line, echo the build commands, otherwise echo
abbreviated commands. Special handling is needed because of the "if"
commands in the build rules have echo unconditionally disabled via @.